requirements.txt
appdirs==1.4.4
asgiref==3.4.1
backcall==0.2.0
backports.entry-points-selectable==1.1.0
black==21.6b0
cairocffi==1.2.0
CairoSVG==2.5.2
cffi==1.14.6
click==8.0.1
colorama==0.4.4
coverage==5.5
cssselect2==0.4.1
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-cleanup==5.2.0
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-import-export==2.5.0
django-markdownx==3.0.1
django-tinymce==3.3.0
et-xmlfile==1.1.0
factory-boy==3.2.0
Faker==8.10.3
filelock==3.0.12
flake8==3.9.2
html5lib==1.1
importlib-metadata==4.10.1
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.1
jedi==0.18.0
Markdown==3.3.6
MarkupPy==1.14
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.0
parso==0.8.2
pathspec==0.9.0
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.1.0
pluggy==0.13.1
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
pyparsing==2.4.7
pyphen==0.11.0
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
text-unidecode==1.3
tinycss2==1.1.0
toml==0.10.2
tox==3.23.1
traitlets==5.0.5
virtualenv==20.6.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
